refactor: ReportEntity 분리
