+++ back-end/legeno-around-here/src/main/java/wooteco/team/ittabi/legenoaroundhere/domain/report/PostSnapshot.java
@@ -1,4 +1,4 @@
-package wooteco.team.ittabi.legenoaroundhere.domain.post;
+package wooteco.team.ittabi.legenoaroundhere.domain.report;
 
 import java.util.ArrayList;
 import java.util.List;
