+++ back-end/legeno-around-here/src/main/java/wooteco/team/ittabi/legenoaroundhere/repository/PostReportRepository.java
@@ -1,7 +1,7 @@
 package wooteco.team.ittabi.legenoaroundhere.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
-import wooteco.team.ittabi.legenoaroundhere.domain.post.PostReport;
+import wooteco.team.ittabi.legenoaroundhere.domain.report.PostReport;
 
 public interface PostReportRepository extends JpaRepository<PostReport, Long> {
