+++ back-end/legeno-around-here/src/main/java/wooteco/team/ittabi/legenoaroundhere/service/ReportService.java
@@ -7,8 +7,8 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import wooteco.team.ittabi.legenoaroundhere.config.IAuthenticationFacade;
 import wooteco.team.ittabi.legenoaroundhere.domain.post.Post;
-import wooteco.team.ittabi.legenoaroundhere.domain.post.PostReport;
-import wooteco.team.ittabi.legenoaroundhere.domain.post.PostSnapshot;
+import wooteco.team.ittabi.legenoaroundhere.domain.report.PostReport;
+import wooteco.team.ittabi.legenoaroundhere.domain.report.PostSnapshot;
 import wooteco.team.ittabi.legenoaroundhere.domain.user.User;
 import wooteco.team.ittabi.legenoaroundhere.dto.PostReportAssembler;
 import wooteco.team.ittabi.legenoaroundhere.dto.PostReportCreateRequest;
